Move UpgradeInfo back to DA.Daml.Options.Types
