+++ sdk/compiler/damlc/daml-package-config/src/DA/Daml/Package/Config.hs
@@ -7,10 +7,6 @@
 module DA.Daml.Package.Config
     ( MultiPackageConfigFields (..)
     , PackageConfigFields (..)
-    , UpgradeInfo (..)
-    , defaultUiTypecheckUpgrades
-    , defaultUiWarnBadInterfaceInstances
-    , defaultUpgradeInfo
     , parseProjectConfig
     , overrideSdkVersion
     , withPackageConfig
@@ -64,23 +60,6 @@ data PackageConfigFields = PackageConfigFields
     , pSdkVersion :: UnresolvedReleaseVersion
     , pUpgradeDar :: Maybe FilePath
     }
-
-data UpgradeInfo = UpgradeInfo
-    { uiUpgradedPackagePath :: Maybe FilePath
-    , uiTypecheckUpgrades :: Bool
-    , uiWarnBadInterfaceInstances :: Bool
-    }
-
-defaultUpgradeInfo :: UpgradeInfo
-defaultUpgradeInfo = UpgradeInfo
-    { uiUpgradedPackagePath = Nothing
-    , uiTypecheckUpgrades = defaultUiTypecheckUpgrades
-    , uiWarnBadInterfaceInstances = defaultUiWarnBadInterfaceInstances
-    }
-
-defaultUiTypecheckUpgrades, defaultUiWarnBadInterfaceInstances :: Bool
-defaultUiTypecheckUpgrades = True
-defaultUiWarnBadInterfaceInstances = False
 
 -- | Parse the daml.yaml for package specific config fields.
 parseProjectConfig :: ProjectConfig -> Either ConfigError PackageConfigFields
