+++ sdk/compiler/damlc/lib/DA/Cli/Damlc.hs
@@ -119,6 +119,7 @@ import DA.Daml.Options.Types (EnableScenarioService(..),
                               Options(Options),
                               SkipScenarioValidation(..),
                               StudioAutorunAllScenarios,
+                              UpgradeInfo(..),
                               damlArtifactDir,
                               distDir,
                               getLogger,
@@ -142,7 +143,6 @@ import DA.Daml.Options.Types (EnableScenarioService(..),
                               projectPackageDatabase)
 import DA.Daml.Package.Config (MultiPackageConfigFields(..),
                                PackageConfigFields(..),
-                               UpgradeInfo(..),
                                checkPkgConfig,
                                findMultiPackageConfig,
                                withPackageConfig,
