+++ sdk/compiler/damlc/tests/src/DA/Test/DamlcIntegration.hs
@@ -82,7 +82,6 @@ import Options.Applicative (execParser, forwardOptions, info, many, strArgument)
 import Outputable (ppr, showSDoc)
 import qualified Proto3.Suite.JSONPB as JSONPB
 import DA.Daml.Project.Types (unsafeResolveReleaseVersion, parseUnresolvedVersion)
-import DA.Daml.Package.Config
 
 import Test.Tasty
 import Test.Tasty.Golden (goldenVsStringDiff)
